Add IRNode
IRNode represents a logical piece of output data. IRNode is either an
instruction (IRInstruction) or a label (IRLabel).

IRInstruction consists of a vector of encoded items. Each encoded item is a
piece of instruction, which is intended to turn into a sequence of bytes later
on.

Change map API: do not copy keys and values, store const pointers to keys and
just pointers to values. Add the destructor in case there is a need to free
values stored in a map.

Store instruction and labels in nodes vector.